# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TODO API",
        "description": "REST API для приложения «Список задач». Авторизация — Bearer-токены (Laravel Sanctum).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Задачи"
                ],
                "summary": "Список задач с фильтрами, сортировкой и пагинацией",
                "description": "Админ видит все задачи; обычный пользователь — только свои.",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "in_progress",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск по заголовку (подстрока)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Поле сортировки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "due_date",
                                "status",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Направление сортировки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Размер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пагинированный список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        },
                                        "links": {
                                            "description": "Ссылки пагинации",
                                            "type": "object"
                                        },
                                        "meta": {
                                            "description": "Метаданные пагинации",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Задачи"
                ],
                "summary": "Создать задачу для текущего пользователя",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Заголовок (3-255 символов)",
                                        "type": "string",
                                        "example": "Купить молоко",
                                        "maxLength": 255,
                                        "minLength": 3
                                    },
                                    "description": {
                                        "description": "Описание",
                                        "type": "string",
                                        "example": "2 литра овсяного",
                                        "nullable": true
                                    },
                                    "due_date": {
                                        "description": "Дедлайн (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-01",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Статус",
                                        "type": "string",
                                        "example": "pending",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The title field must be at least 3 characters."
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "properties": {
                    "id": {
                        "description": "Идентификатор задачи",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Владелец задачи",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Заголовок (3-255 символов)",
                        "type": "string",
                        "example": "Купить молоко"
                    },
                    "description": {
                        "description": "Описание",
                        "type": "string",
                        "example": "2 литра овсяного",
                        "nullable": true
                    },
                    "due_date": {
                        "description": "Дедлайн",
                        "type": "string",
                        "format": "date",
                        "example": "2026-08-01",
                        "nullable": true
                    },
                    "status": {
                        "description": "Статус: pending — ожидающая, in_progress — в работе, completed — завершена",
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed"
                        ]
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    },
                    "can_update": {
                        "description": "Может ли текущий пользователь редактировать",
                        "type": "boolean",
                        "example": true
                    },
                    "can_delete": {
                        "description": "Может ли текущий пользователь удалить",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer-токен Laravel Sanctum. Возвращается из POST /api/auth/login.",
                "scheme": "bearer"
            }
        }
    },
...
}